% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearRegression.R
\name{LinearRegssion}
\alias{LinearRegssion}
\title{Linear Regression}
\usage{
LinearRegssion(theModel, data)
}
\arguments{
\item{theModel}{user-defined model to fit in linear regression}

\item{data}{the data set this linear regression model is fitting on}
}
\value{
a summary of the linear regression model including the residual statistics, coefficients, statistics values (t and F), p-values, and degrees of freedoms
}
\description{
Fit a linear regression model
}
\examples{
LinearRegssion(mpg ~ cyl + disp + hp, data = mtcars)
LinearRegssion(mpg ~ cyl + disp + cyl*disp, data = mtcars)

}
