% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepSelect.R
\name{StepSelect}
\alias{StepSelect}
\title{Step Selection}
\usage{
StepSelect(full, direction, k = 2, trace = FALSE)
}
\arguments{
\item{full}{the full model for backward selection to start and the upper limit for forward selection}

\item{direction}{should be either "backward" or "forward"}

\item{k}{numeric, the penalty per parameter to be used in AIC; the default k = 2 is the classical AIC.}
}
\value{
the final linear regression result after selection
}
\description{
Conduct forward or backward selection for linear regression models
}
\examples{
StepSelect(lm(mpg ~ ., data = mtcars), direction = "forward")
StepSelect(lm(mpg ~ ., data = mtcars), direction = "backward")

}
