% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepSelect.R
\name{StepSelect}
\alias{StepSelect}
\title{Step Selection}
\usage{
StepSelect(full, direction, k = 2, trace = FALSE)
}
\arguments{
\item{full}{the full model for backward selection to start and the upper limit for forward selection}

\item{direction}{should be either "backward" or "forward"}

\item{k}{numeric, the penalty per parameter to be used in AIC; the default k = 2 is the classical AIC.}

\item{trace}{to determine if the process of selection will be shown or not}
}
\value{
the final linear regression result after selection
}
\description{
Conduct forward or backward selection for linear regression models
}
\examples{
full_lm = lm(mpg ~ ., data = mtcars)
n = dim(mtcars)[1]
StepSelect(full_lm, direction = "forward", k = log(n))
StepSelect(full_lm, direction = "backward", trace = TRUE)

}
